refactor: change types for bullet phases

+++ freedroid/src/structs.rs
@@ -182,8 +182,8 @@ pub struct BulletSpec<'sdl> {
     pub recharging_time: f32, // time until the next shot can be made, measures in seconds
     pub speed: f32,           /* speed of the bullet */
     pub damage: i32,          /* damage done by this bullettype */
-    pub phases: i32,          /* how many phases in motion to show */
-    pub phase_changes_per_second: f32, // how many different phases to display every second
+    pub phases: u8,           /* how many phases in motion to show */
+    pub phase_changes_per_second: u16, // how many different phases to display every second
     pub blast: i32,           /* which blast does this bullet create */
     pub surfaces: [Option<Surface<'sdl>>; MAX_PHASES_IN_A_BULLET], // A pointer to the surfaces containing
                                                                    // the bullet images of this bullet
